Initial version from The Gathering 2006.
This was the last day of the TG06 and I had been too late to submit this to the
useless utility contest. My original intent was to create a brainf*ck
interpreter in shell script, which in turn provides a DNS server that always
resolves to localhost and 127.0.0.1.

Unfortunately, I wasted too much time because of that bashism. But maybe
nowadays I could finish it or at least improve it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>