Remove the example/testing brainfuck code.
This was a small debugging snipped I used back then when I was writing
this. In order to be possibly to supply brainfuck code without the need
to change the shell script, the brainfuck code is now accepted as an
argument to the script.

Which was my original intend, because that implementation was just
commented out with the testing code invocation below.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>